sagemaker_batch_transform/tensorflow_cifar-10_with_inference_script/code/requirements.txt
Pillow
numpy
